+++ compiler-cli/src/lib.rs
@@ -448,6 +448,9 @@ enum Hex {
 
     /// Authenticate with Hex
     Authenticate,
+
+    /// Deauthenticate from Hex
+    Deauthenticate,
 }
 
 #[derive(Subcommand, Debug)]
@@ -568,6 +571,8 @@ fn parse_and_run_command() -> Result<(), Error> {
         }
 
         Command::Hex(Hex::Authenticate) => hex::authenticate(),
+
+        Command::Hex(Hex::Deauthenticate) => hex::deauthenticate(),
 
         Command::New(options) => new::create(options, COMPILER_VERSION),